+++ src/main/java/net/spell_engine/compat/TrinketsCompat.java
@@ -5,6 +5,8 @@ import net.fabricmc.fabric.api.util.TriState;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
+import net.minecraft.registry.Registries;
+import net.minecraft.registry.Registry;
 import net.minecraft.util.Identifier;
 import net.spell_engine.SpellEngineMod;
 import net.spell_engine.api.item.trinket.SpellBookTrinketItem;
@@ -33,6 +35,7 @@ public class TrinketsCompat {
                 }
                 return TriState.DEFAULT;
             });
+            Registry.register(Registries.SOUND_EVENT, SpellBookTrinketItem.EQUIP_SOUND_ID, SpellBookTrinketItem.EQUIP_SOUND);
         }
         intialized = true;
     }